refactor: separate emissions folder
