+++ contracts/emissions/interfaces/common/IGaugeUpkeepManager.sol
